fix: FI-Messungen (Stuecknachweis) akzeptieren Dezimalwerte fuer mA/ms
Eingabe von Dezimal- oder Kommawerten in delta_i (mA) / delta_t (ms) loeste
beim Auto-Save einen 500-Fehler aus (int() auf '15.5' -> ValueError), wodurch
die ganze Zeile inkl. delta_t nicht gespeichert wurde.

- models.py: delta_i_ma/delta_t_ms Integer -> Float (SQLite dynamisch typisiert,
  keine Migration noetig)
- stuecknachweis.py: robuster _parse_num()-Helfer (float oder None, wirft nie,
  akzeptiert Komma) ersetzt int() in POST-Handler + Auto-Save
- app.py: Jinja-Filter numfmt (25.0 -> 25, 15.5 bleibt 15.5)
- formular.html / pdf_stuecknachweis.html: step=any an Number-Inputs, Anzeige via numfmt

+++ models.py
@@ -882,8 +882,8 @@ class FiMessung(db.Model):
     sicherung = db.Column(db.String(20), nullable=False)    # z.B. 'F302.2'
     fehlerstrom_30 = db.Column(db.Boolean, default=False)    # 30mA
     fehlerstrom_300 = db.Column(db.Boolean, default=True)    # 300mA
-    delta_i_ma = db.Column(db.Integer, nullable=True)        # ∆I FI [mA]
-    delta_t_ms = db.Column(db.Integer, nullable=True)        # ∆t FI [ms]
+    delta_i_ma = db.Column(db.Float, nullable=True)          # ∆I FI [mA] (Dezimalwerte erlaubt)
+    delta_t_ms = db.Column(db.Float, nullable=True)          # ∆t FI [ms] (Dezimalwerte erlaubt)
     status = db.Column(db.Boolean, default=True)
     reihenfolge = db.Column(db.Integer, default=0)
     manuell = db.Column(db.Boolean, default=False)
